backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
pymongo==4.14.1
pydantic==2.11.7
python-dotenv==1.0.1
numpy==2.1.3
shap==0.46.0
